refactor(uploads): extract publication and resolver owners
